refactor: prepare for multiple k8s RPs
